Organized the code and put comments
