+++ src/app/admin/admin-orders/admin-orders.component.ts
@@ -1,5 +1,4 @@
 import { Component, OnDestroy, ViewEncapsulation } from '@angular/core';
-import { ProductService } from '../../product.service';
 
 @Component({
   selector: 'app-admin-orders',
