+++ src/app/admin/product-form/product-form.component.html
@@ -1,6 +1,11 @@
+<!-- Divide the page into 2 parts: one for the form and another for displaying card-->
+
 <div class="row">
+
   <div class="col-md-6">
+    <!--Product Form-->
     <form #f="ngForm" (ngSubmit)="save(f.value)">
+      <!--Title and its validations-->
       <div class="for-group">
         <label for="title">Title</label>
         <input #title="ngModel" required [(ngModel)]="product.title" name="title" id="title" type="text" class="form-control">
@@ -8,7 +13,9 @@
           Title is required
         </div>
       </div>
+      <!--End of Title-->
 
+      <!--Price and its validations-->
       <div class="form-group">
         <label for="price">Price</label>
         <div class="input-group">
@@ -19,13 +26,15 @@
           <div *ngIf="price.errors.required">
             Price is required
           </div>
-
+          <!--Custome validation using ng2-validation-->
           <div *ngIf="price.errors.min">
             Price must be greater than 0
           </div>
         </div>
       </div>
+      <!--End of Price-->
 
+      <!--Category and its validations-->
       <div class="form-group">
         <label for="category">Category</label>
         <select #category="ngModel" required [(ngModel)]="product.category" name="category" id="category" type="text" class="form-control">
@@ -36,7 +45,9 @@
           Category is required
         </div>
       </div>
+      <!--End of Category-->
 
+      <!--ImageURL and its validations-->
       <div class="form-group">
         <label for="imageUrl">Image URL</label>
         <input url #imageUrl="ngModel" required [(ngModel)]="product.imageUrl" name="imageUrl" id="imageUrl" type="text" class="form-control">
@@ -44,19 +55,25 @@
           <div *ngIf="imageUrl.errors.required">
             Image URL is required
           </div>
-
           <div *ngIf="imageUrl.errors.url">
             Image URL is not valid
           </div>
         </div>
       </div>
+      <!--End of ImageURL-->
 
+      <!--Buttons-->
       <button class="btn btn-primary">Save</button>
-      <button type="button" (click)="delete()" class="btn btn-danger">Delete</button>
-    </form>
 
+      <button type="button" (click)="delete()" class="btn btn-danger">Delete</button>
+
+    </form>
+    <!--End of Product Form-->
   </div>
+
+  <!--Display cards-->
   <div class="col-md-6">
     <product-card [product]="product" [show-actions]="false"></product-card>
   </div>
 </div>
+<!--End of Cards-->
