+++ src/app/bs-navbar/bs-navbar.component.html
@@ -1,5 +1,9 @@
 <nav class="navbar navbar-expand-md navbar-light bg-light fixed-top">
+  <!--Home Button-->
   <a class="navbar-brand" routerLink="/">O</a>
+  <!--End of Home-->
+
+  <!--Dropdown menu-->
   <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarsExampleDefault" aria-controls="navbarsExampleDefault"
     aria-expanded="false" aria-label="Toggle navigation">
     <span class="navbar-toggler-icon"></span>
@@ -7,10 +11,13 @@
 
   <div class="collapse navbar-collapse" id="navbarsExampleDefault">
     <ul class="navbar-nav mr-auto">
+      <!--Shopping Cart link-->
       <li class="nav-item">
         <a class="nav-link" routerLink="/shopping-cart">Shopping Cart</a>
       </li>
+      <!--End of Shopping Cart-->
 
+      <!--This will render either the login button or user's name, depend on if user has logged in or not-->
       <ng-template #anonymousUser>
         <li class="nav-item">
           <a class="nav-link" routerLink="/login">Login</a>
@@ -34,6 +41,8 @@
 
         </div>
       </li>
+      <!--End of user operations-->
     </ul>
   </div>
+  <!--End of dropdown-->
 </nav>
