+++ src/app/bs-navbar/bs-navbar.component.ts
@@ -1,10 +1,9 @@
 import { Observable } from 'rxjs/Rx';
 import { AppUser } from '../model/app-user';
 import { Component, OnInit, ViewEncapsulation } from '@angular/core';
-
-
-import { AuthService } from '../auth.service';
 import { OnDestroy } from '@angular/core/src/metadata/lifecycle_hooks';
+import { AuthService } from "../services/auth-services/auth.service";
+
 @Component({
   selector: 'bs-navbar',
   templateUrl: './bs-navbar.component.html',
@@ -12,19 +11,24 @@ import { OnDestroy } from '@angular/core/src/metadata/lifecycle_hooks';
   encapsulation: ViewEncapsulation.None
 })
 export class BsNavbarComponent implements OnDestroy {
-  ngOnDestroy(): void {
-    this.subscription.unsubscribe();
-  }
-
+  // hold the user info
   appUser: AppUser;
+
   subscription;
+
   constructor(private auth: AuthService) {
+    // get the app user by AuthService
     this.subscription = auth.appUser$.subscribe(appUser => this.appUser = appUser);
   }
 
+  // log out
   logout() {
     this.auth.logout();
   }
 
+  // unsubscribe
+  ngOnDestroy(): void {
+    this.subscription.unsubscribe();
+  }
 
 }
