+++ src/app/login/login.component.html
@@ -1,3 +1,5 @@
+<!--just a log in button with a click event-->
+
 <button (click)="login()" class="btn btn-primary">
   Login with Googe
 </button>
