+++ src/app/login/login.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit, ViewEncapsulation } from '@angular/core';
-import { AuthService } from '../auth.service';
+import { AuthService } from "../services/auth-services/auth.service";
 
 
 @Component({
@@ -12,6 +12,7 @@ export class LoginComponent {
 
   constructor(private auth: AuthService) { }
 
+  // log in method by AuthService
   login() {
     this.auth.login();
   }
