+++ src/app/model/app-user.ts
@@ -1,3 +1,4 @@
+// this interface defines user info 
 export interface AppUser {
     name: string;
     email: string;
