+++ src/app/model/product.ts
@@ -1,3 +1,4 @@
+// defines product's info
 export interface Product {
     title: string;
     price: number;
