+++ src/app/product-card/product-card.component.html
@@ -1,3 +1,5 @@
+<!--Display the card's info-->
+
 <div *ngIf="product.title" class="card h-100">
   <img class="card-img-top" [src]="product.imageUrl" *ngIf="product.imageUrl">
   <div class="card-body">
