+++ src/app/products/product-filter/product-filter.component.html
@@ -1,3 +1,4 @@
+<!--just display a list of categories and highlight the selected one-->
 <div class="sticky-top">
   <div class="list-group">
     <a class="list-group-item list-group-item-action" [class.active]="!category" routerLink="/">All Categories</a>
