+++ src/app/products/product-filter/product-filter.component.ts
@@ -1,5 +1,5 @@
-import { CategoryService } from '../../category.service';
-import { Component, Input, OnInit, ViewEncapsulation } from '@angular/core';
+import { Component, Input, ViewEncapsulation } from '@angular/core';
+import { CategoryService } from "../../services/goods-services/category.service";
 
 @Component({
   selector: 'product-filter',
@@ -7,14 +7,13 @@ import { Component, Input, OnInit, ViewEncapsulation } from '@angular/core';
   styleUrls: ['./product-filter.component.css'],
   encapsulation: ViewEncapsulation.None
 })
-export class ProductFilterComponent implements OnInit {
+export class ProductFilterComponent {
+  // the list of categories from the database
   categories$;
+  // this input to high light the current selected category
   @Input('category') category;
   constructor(categoryService: CategoryService) {
+    // get the categories from the database by CategoryService
     this.categories$ = categoryService.getAll();
   }
-
-  ngOnInit() {
-  }
-
 }
