+++ src/app/services/auth-services/auth.service.ts
@@ -1,5 +1,4 @@
-import { UserService } from './user.service';
-import { AppUser } from './model/app-user';
+
 import { ActivatedRoute } from '@angular/router';
 import { Observable } from 'rxjs/Rx';
 import { AngularFireAuth } from 'angularfire2/auth';
@@ -7,6 +6,8 @@ import { Injectable } from '@angular/core';
 import * as firebase from 'firebase';
 import 'rxjs/add/operator/switchMap';
 import 'rxjs/add/observable/of'
+import { UserService } from "../user-services/user.service";
+import { AppUser } from "../../model/app-user";
 @Injectable()
 export class AuthService {
